refactor(server): :truck: restructuring
Restructuring the server files. Cleaning up index.ts.
