+++ src/core/game/player.ts
