+++ src/core/game/wordgame.ts
